Refactor: split journey UI logic & extract shared header/footer components
